requirements.txt
sounddevice>=0.5.2
scipy>=1.9.3
transformers>=4.40.1
qdrant-client>=1.8.0
python-dotenv>=1.1.0
sentence-transformers>=2.7.0
accelerate>=0.26.0
torch>=2.1.0
fastapi>=0.110.0
uvicorn>=0.29.0
